Organize endpoint and field related widgets
New mods are about to be added to split the EndpointPane in additional
panes. Before starting this, I'm adding a new parent module in the
hierarchy to separate the molecule widgets related to the EndpointPane.
