+++ src/widgets/endpoint/mod.rs
@@ -15,27 +15,10 @@
 //
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-mod authorization_pane;
-mod code_view;
-pub mod dialogs;
-pub mod endpoint;
-mod error_pane;
-mod export_dialog;
-pub mod field;
-mod file_dialogs;
-mod key_value_pane;
-mod key_value_row;
-mod method_dropdown;
-mod request_body;
-mod search_box;
+mod endpoint_pane;
+mod response_headers;
+mod response_panel;
 
-pub use authorization_pane::*;
-pub use code_view::CodeView;
-pub use error_pane::ErrorPane;
-pub use export_dialog::*;
-pub use file_dialogs::*;
-pub use key_value_pane::KeyValuePane;
-pub use key_value_row::KeyValueRow;
-pub use method_dropdown::MethodDropdown;
-pub use request_body::*;
-pub use search_box::SearchBox;
+pub use endpoint_pane::*;
+pub use response_headers::*;
+pub use response_panel::*;
